添加订单接口

+++ sql/sfm_credit_card.sql
@@ -24,8 +24,8 @@ DROP TABLE IF EXISTS `sfm_credit_card`;
 CREATE TABLE `sfm_credit_card` (
   `id` int(10) unsigned NOT NULL PRIMARY KEY AUTO_INCREMENT COMMENT '自增id',
   `card_id` varchar(32) COMMENT '额度卡编号',
-  `amount` decimal COMMENT '总额度',
-  `remain_amount` decimal COMMENT '剩余额度',
+  `amount` int(10) COMMENT '总额度',
+  `remain_amount` int(10) COMMENT '剩余额度',
   INDEX `card_id` (`card_id`) COMMENT '额度卡编号索引'
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COMMENT='信用额度卡表';
